Add initial work on wg-activate command
Currently, this only allows one peer per interface and expects the
interface to be configured as pointtopoint. In the future, the plan is
to add IPAM so we don't need to be so strict about the interface
configuration.

There is also currently no support for setting MTU despite the option
being available in the client config JSON.